削除機能実装

+++ AutoSystem_springboot/src/main/java/com/example/repository/AnswersRepository.java
@@ -9,4 +9,4 @@ import com.example.entity.AnswersEntity;
  * 答え Repository
  */
 @Repository
-public interface AnswersRepository extends JpaRepository<AnswersEntity,Long> {}
+public interface AnswersRepository extends JpaRepository<AnswersEntity,Integer> {}
