+++ AutoSystem_springboot/src/main/java/com/example/repository/QuestionsRepository.java
@@ -9,4 +9,4 @@ import com.example.entity.QuestionsEntity;
  * 質問 Repository
  */
 @Repository
-public interface QuestionsRepository extends JpaRepository<QuestionsEntity,Long> {}
+public interface QuestionsRepository extends JpaRepository<QuestionsEntity,Integer> {}
